refactor(dispatch): extract Hungarian assignment + scratch into assignment.rs
Pure file-move refactor — no behavior change. The Hungarian solver
(assign / assign_with_scratch), per-pass scratch buffers
(DispatchScratch), and the helpers feeding them
(pending_stops_minus_covered, scale_cost, ASSIGNMENT_* constants)
were a self-contained ~390-line block at the bottom of
dispatch/mod.rs. Lifting them into a dedicated module shrinks
mod.rs (1635 → 1242 LOC) and surfaces the ASSIGNMENT_SENTINEL /
ASSIGNMENT_SCALE constants alongside the math that uses them.

The assignment module is pub(crate); items inside are pub so the
mod.rs re-exports stay simple. AssignmentResult is pub since
external strategies may want to construct it in tests.

Public API surface preserved via pub use re-exports —
crate::dispatch::AssignmentResult / DispatchScratch / assign /
assign_with_scratch all work unchanged.

Continuation of §20 from the audit; manifest.rs landed in #773,
this is the second of three planned splits.
